Fix clang build

+++ src/pc/psxsdk/emu.cpp
@@ -127,7 +127,7 @@ extern "C" u16 read_16(u32 addr, char* file, int line)
 
 extern "C" void write_dma(u32 data, char* file, int line)
 {
-    writes.push_back({SPU->RWAddr, data, file, line, 1});
+    writes.push_back({SPU->RWAddr, (u16)data, file, line, 1});
     SPU->WriteDMA(data);
 }
